Move Settings to config thread
