+++ sumobots/driverstation/src/main/java/org/inspirerobotics/sumobots/driverstation/config/Settings.java
@@ -1,4 +1,4 @@
-package org.inspirerobotics.sumobots.driverstation;
+package org.inspirerobotics.sumobots.driverstation.config;
 
 import java.util.logging.Level;
